requirements.txt
Django==3.1.7
psycopg2==2.8.6
Pillow==8.1.2
python-decouple==3.4
djangorestframework==3.12.2
django-filter==2.4.0
django-cors-headers==3.7.0
gunicorn==20.0.4
django-cleanup==5.1.0
